fix: full network name in buy/swap tab tooltips
