+++ components/brave_wallet_ui/components/buy-send-swap/buy-send-swap-tooltip/index.tsx
@@ -1,5 +1,11 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
 import * as React from 'react'
 
+// style
 import {
   StyledWrapper,
   Tip,
@@ -10,10 +16,11 @@ export interface Props {
   children?: React.ReactNode
   text: string
   isDisabled: boolean
+  maxTextWidth?: React.CSSProperties['maxWidth']
 }
 
 function BuySendSwapTooltip (props: Props) {
-  const { children, text, isDisabled } = props
+  const { children, text, isDisabled, maxTextWidth } = props
   const [active, setActive] = React.useState(false)
 
   const showTip = () => {
@@ -33,7 +40,7 @@ function BuySendSwapTooltip (props: Props) {
       {active && isDisabled && (
         <>
           <Pointer />
-          <Tip>
+          <Tip maxWidth={maxTextWidth}>
             {text}
           </Tip>
         </>
